adds statuses summaries for assemblies+runs to study admin page. refactors admin site.
