+++ templates/admin/study_admin_run_type_summary.html
@@ -1,0 +1,9 @@
+{% extends "unfold/layouts/base.html" %}
+{% load unfold %}
+
+{% block content %}
+  {% component "unfold/components/card.html" %}
+    {% component "unfold/components/table.html" with table=run_types_table card_included=1 striped=1 %}{% endcomponent %}
+  {% endcomponent %}
+
+{% endblock %}
