+++ templates/admin/tabular_inline_paginated_tabbed.html
@@ -1,0 +1,8 @@
+{% include 'admin/edit_inline/tabular.html' %}
+<div
+  class="admin-tabular-inline-pagination js-inline-admin-formset inline-group"
+  id="{{ inline_admin_formset.formset.prefix }}-group-pag"
+  {% if inline_admin_formset.opts.tab %}x-show="activeTab == '{{ inline_admin_formset.opts.verbose_name|slugify }}'"{% else %}x-show="activeTab == 'general'"{% endif %}
+>
+    {% include 'admin/tabular_paginator.html' %}
+</div>
